+++ tests/Utils/SchemaPrinterTest.php
@@ -17,15 +17,20 @@ use GraphQL\Utils\SchemaPrinter;
 use PHPUnit\Framework\TestCase;
 
 /**
+ * @phpstan-import-type Options from SchemaPrinter
+ *
  * @see describe('Type System Printer', () => {
  */
 final class SchemaPrinterTest extends TestCase
 {
-    private static function assertPrintedSchemaEquals(string $expected, Schema $schema): void
+    /**
+     * @param Options $options
+     */
+    private static function assertPrintedSchemaEquals(string $expected, Schema $schema, array $options = []): void
     {
-        $printedSchema = SchemaPrinter::doPrint($schema);
+        $printedSchema = SchemaPrinter::doPrint($schema, $options);
         $builtSchema = BuildSchema::build($printedSchema);
-        $cycledSchema = SchemaPrinter::doPrint($builtSchema);
+        $cycledSchema = SchemaPrinter::doPrint($builtSchema, $options);
 
         self::assertEquals($printedSchema, $cycledSchema);
         self::assertEquals($expected, $printedSchema);
@@ -1169,5 +1174,38 @@ final class SchemaPrinterTest extends TestCase
 
       GRAPHQL;
         self::assertEquals($expected, $output);
+    }
+
+    /**
+     * Additional functionality not present in the reference implementation.
+     */
+    public function testPrintSchemaWithSortedTypes(): void
+    {
+        $schema = new Schema([
+            'types' => [
+                new EnumType(['name' => 'SomeEnum', 'values' => []]),
+                new InputObjectType(['name' => 'InputObject', 'fields' => []]),
+                new InterfaceType(['name' => 'FooInterface', 'fields' => []]),
+                new ObjectType(['name' => 'Abc', 'fields' => []]),
+                new UnionType(['name' => 'Union', 'types' => []]),
+            ],
+        ]);
+
+        self::assertPrintedSchemaEquals(
+            <<<'GRAPHQL'
+            type Abc
+
+            interface FooInterface
+
+            input InputObject
+            
+            enum SomeEnum
+
+            union Union
+
+            GRAPHQL,
+            $schema,
+            ['sortTypes' => true]
+        );
     }
 }